src/user_service/requirements.txt
bcrypt==4.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
email_validator==2.2.0
Flask==3.0.3
pydantic==2.9.2
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.6
PyJWT==2.9.0
cryptography==43.0.3
cffi==1.17.1
pycparser==2.22


